\name{predict_m6a_benifit}
\alias{hello}
\title{Hello, World!}
\usage{
predict_m6a_benifit(data,cancer)
}
\description{
The data must be a data frame for age, stage, grade, and Entrez expression profile data, which can be customized based on the sample data; cancer is a string, one of ACC, BLCA, BRCA, CESC, COAD, ESCA, GBM, HNSC, KIRC, KIRP, LGG, LIHC, LUAD, LUSC, MESO, OV, PAAD, PCPG, PRAD, READ, SARC, SKCM, STAD, TGCT, THCA, THYM, UCEC or UVM.
}
\examples{

}


